Move renameTypeParameter to SuggestedFixes.
It's shared by three checks; this seems like a natural home.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=291343102
