+++ check_api/src/main/java/com/google/errorprone/fixes/SuggestedFixes.java
@@ -33,6 +33,7 @@ import static java.util.stream.Collectors.joining;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
+import com.google.common.base.Objects;
 import com.google.common.base.Predicates;
 import com.google.common.base.Verify;
 import com.google.common.collect.FluentIterable;
@@ -50,10 +51,12 @@ import com.google.errorprone.util.ASTHelpers;
 import com.google.errorprone.util.ErrorProneToken;
 import com.google.errorprone.util.FindIdentifiers;
 import com.sun.source.doctree.DocTree;
+import com.sun.source.doctree.ParamTree;
 import com.sun.source.tree.AnnotationTree;
 import com.sun.source.tree.AssignmentTree;
 import com.sun.source.tree.BinaryTree;
 import com.sun.source.tree.ClassTree;
+import com.sun.source.tree.CompilationUnitTree;
 import com.sun.source.tree.CompoundAssignmentTree;
 import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.IdentifierTree;
@@ -64,8 +67,11 @@ import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.ModifiersTree;
 import com.sun.source.tree.NewArrayTree;
 import com.sun.source.tree.Tree;
+import com.sun.source.tree.TypeParameterTree;
 import com.sun.source.tree.VariableTree;
+import com.sun.source.util.DocSourcePositions;
 import com.sun.source.util.DocTreePath;
+import com.sun.source.util.DocTreeScanner;
 import com.sun.source.util.JavacTask;
 import com.sun.source.util.TreePath;
 import com.sun.tools.javac.api.BasicJavacTask;
@@ -83,6 +89,7 @@ import com.sun.tools.javac.parser.Tokens;
 import com.sun.tools.javac.parser.Tokens.Comment;
 import com.sun.tools.javac.parser.Tokens.TokenKind;
 import com.sun.tools.javac.tree.DCTree;
+import com.sun.tools.javac.tree.DCTree.DCDocComment;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.JCTree.JCCompilationUnit;
 import com.sun.tools.javac.tree.JCTree.JCIdent;
@@ -636,6 +643,77 @@ public class SuggestedFixes {
       MethodInvocationTree tree) {
     return new IllegalStateException(
         String.format("Couldn't replace the method name in %s.", tree));
+  }
+
+  /**
+   * Renames a type parameter {@code typeParameter} owned by {@code owningTree} to {@code
+   * typeVarReplacement}. Renames occurrences in Javadoc as well.
+   */
+  public static SuggestedFix renameTypeParameter(
+      TypeParameterTree typeParameter,
+      Tree owningTree,
+      String typeVarReplacement,
+      VisitorState state) {
+    Symbol typeParameterSymbol = getSymbol(typeParameter);
+
+    // replace only the type parameter name (and not any upper bounds)
+    String name = typeParameter.getName().toString();
+    int pos = ((JCTree) typeParameter).getStartPosition();
+    Builder fixBuilder =
+        SuggestedFix.builder().replace(pos, pos + name.length(), typeVarReplacement);
+
+    ((JCTree) owningTree)
+        .accept(
+            new TreeScanner() {
+              @Override
+              public void visitIdent(JCIdent tree) {
+                Symbol identSym = getSymbol(tree);
+                if (Objects.equal(identSym, typeParameterSymbol)) {
+                  // Lambda parameters can be desugared early, so we need to make sure the source
+                  // is there. In the example below, we would try to suggest replacing the node 't'
+                  // with T2, since the compiler desugars to g((T t) -> false). The extra condition
+                  // prevents us from doing that.
+
+                  // Foo<T> {
+                  //   <G> void g(Predicate<G> p) {},
+                  //   <T> void blah() {
+                  //     g(t -> false);
+                  //   }
+                  // }
+                  if (Objects.equal(state.getSourceForNode(tree), name)) {
+                    fixBuilder.replace(tree, typeVarReplacement);
+                  }
+                }
+              }
+            });
+    DCDocComment docCommentTree =
+        (DCDocComment) JavacTrees.instance(state.context).getDocCommentTree(state.getPath());
+    if (docCommentTree != null) {
+      docCommentTree.accept(
+          new DocTreeScanner<Void, Void>() {
+            @Override
+            public Void visitParam(ParamTree paramTree, Void unused) {
+              if (paramTree.isTypeParameter()
+                  && paramTree.getName().getName().contentEquals(name)) {
+                DocSourcePositions positions =
+                    JavacTrees.instance(state.context).getSourcePositions();
+                CompilationUnitTree compilationUnitTree = state.getPath().getCompilationUnit();
+                int startPos =
+                    (int)
+                        positions.getStartPosition(
+                            compilationUnitTree, docCommentTree, paramTree.getName());
+                int endPos =
+                    (int)
+                        positions.getEndPosition(
+                            compilationUnitTree, docCommentTree, paramTree.getName());
+                fixBuilder.replace(startPos, endPos, typeVarReplacement);
+              }
+              return super.visitParam(paramTree, null);
+            }
+          },
+          null);
+    }
+    return fixBuilder.build();
   }
 
   /** Deletes the given exceptions from a method's throws clause. */
